auction-backend/requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-celery-results==2.0.0
django-cors-headers==3.6.0
django-extensions==3.1.0
django-utils-six==2.0
djangorestframework==3.12.2
importlib-metadata==3.3.0
kombu==5.0.2
Pillow==8.0.1
prompt-toolkit==3.0.10
psycopg2==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-celery-results==2.0.0
django-cors-headers==3.6.0
django-extensions==3.1.0
django-utils-six==2.0
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.3.0
kombu==5.0.2
Pillow==8.0.1
prompt-toolkit==3.0.10
psycopg2==2.8.6
python-firebase==1.2
pytz==2020.4
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.2.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-celery-results==2.0.0
django-cors-headers==3.6.0
django-extensions==3.1.0
django-utils-six==2.0
djangorestframework==3.12.2
firebase-admin==4.5.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
httplib2==0.18.1
idna==2.10
importlib-metadata==3.3.0
kombu==5.0.2
msgpack==1.0.2
Pillow==8.0.1
prompt-toolkit==3.0.10
proto-plus==1.13.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-firebase==1.2
pytz==2020.4
redis==3.5.3
requests==2.25.1
rsa==4.7
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
